move combine-related functions into convert_combine.py to declutter
